use the data in a route to just pass if something i a popup
