+++ apps/servicebus-browser-frontend/src/app/app.component.html
@@ -1,5 +1,1 @@
-<lib-main-ui
-  [electronWindowControlSpacing]="windowControlSpacing()"
-  [useLowProfileMenu]="isMac"
-  [menuItems]="menuItems"
-/>
+<router-outlet />
