+++ apps/servicebus-browser-frontend/src/app/main-shell/main-shell.html
@@ -1,0 +1,5 @@
+<lib-main-ui
+  [electronWindowControlSpacing]="windowControlSpacing()"
+  [useLowProfileMenu]="isMac"
+  [menuItems]="menuItems"
+/>
