+++ libs/messages/flow/src/index.ts
@@ -30,3 +30,10 @@ export const routes = (options: { baseRoute: string }): Routes => [
     ]
   }
 ];
+
+export const popups: Routes = [
+  {
+    path: 'body-viewer/:pageId/:messageKey',
+    loadComponent: () => import('./lib/body-viewer/body-viewer').then(m => m.BodyViewer),
+  }
+];
